Create edit_tables
This file is used to import all the necessary tables in R and edit them, so this won't have to be done in the other scripts. However this script is not complete and it is a bit time-consuming to run. 
It might be nice to rewrite this into functions or something like that, to optimise the runtime.